requirements.txt
torch 
torchvision
streamlit
opencv-python
numpy
pandas
grad-cam
matplotlib
fpdf2